Revert "Revert "Merge branch 'master' of https://github.com/Eytins/HotelManagement""
This reverts commit 5011b3710f56d9ce33f1e774cfccac658d36a519.

+++ src/com/HotelManagement/service/hotelType/HotelTypeServiceImpl.java
@@ -44,7 +44,7 @@ public class HotelTypeServiceImpl implements HotelTypeService {
     }
 
     @Override
-    public int modifyRoomType(Integer roomTypeId, Integer price, Integer numberOfPerson) {
+    public int modifyRoomType(Integer roomTypeId, long price, Integer numberOfPerson) {
         return this.hotelTypeMapper.modifyRoomType(roomTypeId,price,numberOfPerson);
     }
